cambios en el catálogo de productos, cambios en el catalogo de laboratorios, se generó el sp para registrar nuevos laboratorios, ademas se hizo cambio en el modelo de este último porque difería de lo que ya se tenía; cambiaron los nombres de las propiedades...

+++ apiQuiroga.Models/Catalogos/FormulasModel.cs
@@ -8,6 +8,6 @@
       public bool FormulaCombinada { get; set; }
       public string FormulaAccionTerapeutica { get; set; }
       public string FormulaProdLider { get; set; }
-      public short FormulaGenericoID { get; set; }
+      public int FormulaGenericoID { get; set; }
    }
 }
